simulator/requirements.txt
requests
redis
pyyaml
